third_party/blink/web_tests/external/wpt/webnn/idlharness.https.any.worker-expected.txt

This is a testharness.js-based test.
Found 5 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] idl_test setup
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Not implemented"
[FAIL] MLGraph must be primary interface of graph
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: graph is not defined"
[FAIL] Stringification of graph
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: graph is not defined"
[FAIL] MLGraphBuilder interface: operation clamp(MLOperand, optional MLClampOptions)
  assert_equals: property has wrong .length expected 0 but got 1
[FAIL] MLGraphBuilder interface: operation clamp(optional MLClampOptions)
  assert_equals: property has wrong .length expected 0 but got 1
Harness: the test ran to completion.
